Changed model

+++ Joobie/Joobie/Models/JobModels/Company.cs
@@ -2,7 +2,7 @@
 {
     public class Company
     {
-        public byte Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
     }
 }
